+++ source/extensions/isaacsim.ros2.urdf/isaacsim/ros2/urdf/commands.py
@@ -1,0 +1,91 @@
+import asyncio
+import os
+from functools import partial
+
+import omni.client
+import omni.kit.commands
+from isaacsim.ros2.urdf.RobotDescription import RobotDefinitionReader
+
+# import omni.kit.utils
+from omni.client._omniclient import Result
+from omni.importer.urdf import _urdf
+from pxr import Usd, UsdUtils
+
+
+class URDFImportFromROS2Node(omni.kit.commands.Command):
+    def __init__(
+        self,
+        ros2_node_name: str = "robot_state_publisher",
+        import_config=_urdf.ImportConfig(),
+        dest_path: str = "",
+        get_articulation_root: bool = False,
+    ):
+        self.ros2_node_name = ros2_node_name
+        self.dest_path = dest_path
+        self.config = import_config
+        self.robot_definition = RobotDefinitionReader()
+        self.robot_definition.description_received_fn = partial(self.on_description_received)
+        self.robot_model = None
+        self.finished = False
+        self.__subscription = (
+            omni.kit.app.get_app().get_update_event_stream().create_subscription_to_pop(self.on_app_update)
+        )
+
+    def on_app_update(self, event):
+        if self.finished:
+            self.__subscription.unsubscribe()
+            if self.robot_model:
+                self.import_robot(self.robot_model)
+            return
+
+    def on_description_received(self, urdf_description):
+
+        result, robot_model = omni.kit.commands.execute(
+            "URDFParseText", urdf_string=urdf_description, import_config=self.config
+        )
+        if result:
+            self.finished = True
+            self.robot_model = robot_model
+
+    def import_robot(self, robot_model):
+        if self.dest_path == "":
+            all_cache_stage = UsdUtils.StageCache.Get().GetAllStages()
+            if len(all_cache_stage) == 1:
+                result = omni.kit.commands.execute(
+                    "URDFImportRobot",
+                    urdf_robot=robot_model,
+                    import_config=self.config,
+                    dest_path=self.dest_path,
+                )
+                return all_cache_stage[0].GetRootLayer().identifier
+        else:
+            result = omni.kit.commands.execute(
+                "URDFImportRobot",
+                urdf_robot=robot_model,
+                import_config=self.config,
+                dest_path=self.dest_path,
+            )
+            stage = Usd.Stage.Open(self.dest_path)
+            prim_name = str(stage.GetDefaultPrim().GetName())
+
+            # print(prim_name)
+            # stage.Save()
+            def add_reference_to_stage():
+                current_stage = omni.usd.get_context().get_stage()
+                if current_stage:
+                    prim_path = omni.usd.get_stage_next_free_path(
+                        current_stage, str(current_stage.GetDefaultPrim().GetPath()) + "/" + prim_name, False
+                    )
+                    robot_prim = current_stage.OverridePrim(prim_path)
+                    if "anon:" in current_stage.GetRootLayer().identifier:
+                        robot_prim.GetReferences().AddReference(self.dest_path)
+                    else:
+                        robot_prim.GetReferences().AddReference(
+                            omni.client.make_relative_url(current_stage.GetRootLayer().identifier, self.dest_path)
+                        )
+
+            add_reference_to_stage()
+
+    def do(self) -> Result:
+
+        self.robot_definition.start_get_robot_description(self.ros2_node_name)