+++ source/extensions/isaacsim.ros2.urdf/docs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+
+## [1.1.0] - 2024-10-15
+### Changed
+- Moved ros2 node import workflow UI to File->Menu to align with URDF Importer UI
+### Added
+- Create `URDFImportFromROS2Node` command
+
+
 ## [1.0.0] - 2024-09-27
 ### Changed
 - Extension renamed to isaacsim.ros2.urdf